erste Struktur fürs MVC erstellt und FirstFragment und SecondFragment entfernt
