+++ app/src/main/java/com/example/rezepteapp/model/ReciepeModel.java
@@ -1,0 +1,4 @@
+package com.example.rezepteapp.model;
+
+public class ReciepeModel {
+}
