+++ app/src/main/res/layout/fragment_welcome_screen.xml
@@ -5,7 +5,7 @@
     android:id="@+id/layout_frame"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WelcomeScreenFragment">
+    tools:context=".controller.WelcomeScreenFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_constraint"
