+++ app/src/main/res/navigation/nav_graph.xml
@@ -2,32 +2,11 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    android:id="@+id/nav_graph">
 
-    <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.example.rezepteapp.FirstFragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
-
-        <action
-            android:id="@+id/action_FirstFragment_to_entranceFragment"
-            app:destination="@id/entranceFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.example.rezepteapp.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
     <fragment
         android:id="@+id/entranceFragment"
-        android:name="com.example.rezepteapp.WelcomeScreenFragment"
+        android:name="com.example.rezepteapp.controller.WelcomeScreenFragment"
         android:label="fragment_entrance"
         tools:layout="@layout/fragment_welcome_screen" />
 </navigation>
